perf: add light sdk surfaces for ui consumers
