+++ typescript/sdk/src/core/adapters/SealevelCoreAdapter.ts
@@ -21,7 +21,7 @@ import {
   SealevelMailboxTransferOwnershipInstruction,
   SealevelMailboxTransferOwnershipInstructionSchema,
 } from '../../mailbox/serialization.js';
-import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../../providers/ConfiguredMultiProtocolProvider.js';
 import {
   ProviderType,
   TypedTransactionReceipt,
