+++ typescript/sdk/src/ism/adapters/SealevelMultisigAdapter.ts
@@ -16,7 +16,7 @@ import {
   SealevelMultisigIsmTransferOwnershipInstruction,
   SealevelMultisigIsmTransferOwnershipInstructionSchema,
 } from '../../ism/serialization.js';
-import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../../providers/ConfiguredMultiProtocolProvider.js';
 import { ChainName } from '../../types.js';
 import { SealevelInstructionWrapper } from '../../utils/sealevelSerialization.js';
