+++ typescript/sdk/src/providers/providerBuilders.ts
@@ -1,225 +1,50 @@
-import { CosmWasmClient } from '@cosmjs/cosmwasm-stargate';
-import { StargateClient } from '@cosmjs/stargate';
-import { Connection } from '@solana/web3.js';
-import { providers } from 'ethers';
-import { RpcProvider as StarknetRpcProvider } from 'starknet';
-import { createPublicClient, http } from 'viem';
-import { Provider as ZKProvider } from 'zksync-ethers';
+import { ProtocolType } from '@hyperlane-xyz/utils';
 
-import { AleoProvider as AleoSDKProvider } from '@hyperlane-xyz/aleo-sdk/runtime';
-import { CosmosNativeProvider } from '@hyperlane-xyz/cosmos-sdk/runtime';
-import { RadixProvider as RadixSDKProvider } from '@hyperlane-xyz/radix-sdk/runtime';
-import { TronJsonRpcProvider } from '@hyperlane-xyz/tron-sdk/runtime';
-import { ProtocolType, assert, isNumeric } from '@hyperlane-xyz/utils';
+import type { KnownProtocolType, TypedProvider } from './ProviderType.js';
+import { ProviderType } from './ProviderType.js';
 
-import { ChainMetadata, RpcUrl } from '../metadata/chainMetadataTypes.js';
+export type {
+  ProviderBuilderFn,
+  TypedProviderBuilderFn,
+} from './builders/types.js';
+export { defaultAleoProviderBuilder } from './builders/aleo.js';
+export {
+  defaultCosmJsNativeProviderBuilder,
+  defaultCosmJsProviderBuilder,
+  defaultCosmJsWasmProviderBuilder,
+} from './builders/cosmos.js';
+export {
+  defaultEthersV5ProviderBuilder,
+  defaultFuelProviderBuilder,
+  defaultProviderBuilder,
+} from './builders/ethersV5.js';
+export { defaultRadixProviderBuilder } from './builders/radix.js';
+export { defaultSolProviderBuilder } from './builders/solana.js';
+export { defaultStarknetJsProviderBuilder } from './builders/starknet.js';
+export {
+  defaultTronEthersProviderBuilder,
+  defaultTronProviderBuilder,
+} from './builders/tron.js';
+export { defaultViemProviderBuilder } from './builders/viem.js';
+export {
+  defaultZKProviderBuilder,
+  defaultZKSyncProviderBuilder,
+} from './builders/zksync.js';
 
+import type { ProviderBuilderFn } from './builders/types.js';
+import { defaultAleoProviderBuilder } from './builders/aleo.js';
 import {
-  AleoProvider,
-  CosmJsNativeProvider,
-  CosmJsProvider,
-  CosmJsWasmProvider,
-  EthersV5Provider,
-  KnownProtocolType,
-  ProviderType,
-  RadixProvider,
-  SolanaWeb3Provider,
-  StarknetJsProvider,
-  TronProvider,
-  TypedProvider,
-  ViemProvider,
-  ZKSyncProvider,
-} from './ProviderType.js';
-import { HyperlaneSmartProvider } from './SmartProvider/SmartProvider.js';
-import { ProviderRetryOptions } from './SmartProvider/types.js';
-import { parseCustomRpcHeaders } from '../utils/provider.js';
-
-export type ProviderBuilderFn<P> = (
-  rpcUrls: ChainMetadata['rpcUrls'],
-  network: number | string,
-  retryOverride?: ProviderRetryOptions,
-) => P;
-export type TypedProviderBuilderFn = ProviderBuilderFn<TypedProvider>;
-
-const DEFAULT_RETRY_OPTIONS: ProviderRetryOptions = {
-  maxRetries: 3,
-  baseRetryDelayMs: 250,
-};
-
-export function defaultEthersV5ProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: number | string,
-  retryOverride?: ProviderRetryOptions,
-): EthersV5Provider {
-  const provider = new HyperlaneSmartProvider(
-    network,
-    rpcUrls,
-    undefined,
-    retryOverride || DEFAULT_RETRY_OPTIONS,
-  );
-  return { type: ProviderType.EthersV5, provider };
-}
-
-export function defaultViemProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: number | string,
-): ViemProvider {
-  if (!rpcUrls.length) throw new Error('No RPC URLs provided');
-  if (!isNumeric(network)) throw new Error('Viem requires a numeric network');
-  const id = parseInt(network.toString(), 10);
-  const name = network.toString(); // TODO get more descriptive name
-  const url = rpcUrls[0].http;
-  const client = createPublicClient({
-    chain: {
-      id,
-      name,
-      network: name,
-      nativeCurrency: { name: '', symbol: '', decimals: 0 },
-      rpcUrls: { default: { http: [url] }, public: { http: [url] } },
-    },
-    transport: http(rpcUrls[0].http),
-  });
-  return { type: ProviderType.Viem, provider: client };
-}
-
-export function defaultSolProviderBuilder(
-  rpcUrls: RpcUrl[],
-  _network: number | string,
-): SolanaWeb3Provider {
-  if (!rpcUrls.length) throw new Error('No RPC URLs provided');
-  return {
-    type: ProviderType.SolanaWeb3,
-    provider: new Connection(rpcUrls[0].http, 'confirmed'),
-  };
-}
-
-export function defaultFuelProviderBuilder(
-  rpcUrls: RpcUrl[],
-  _network: number | string,
-): EthersV5Provider {
-  if (!rpcUrls.length) throw new Error('No RPC URLs provided');
-  throw new Error('TODO fuel support');
-}
-
-export function defaultCosmJsProviderBuilder(
-  rpcUrls: RpcUrl[],
-  _network: number | string,
-): CosmJsProvider {
-  if (!rpcUrls.length) throw new Error('No RPC URLs provided');
-  return {
-    type: ProviderType.CosmJs,
-    provider: StargateClient.connect(rpcUrls[0].http),
-  };
-}
-
-export function defaultCosmJsWasmProviderBuilder(
-  rpcUrls: RpcUrl[],
-  _network: number | string,
-): CosmJsWasmProvider {
-  if (!rpcUrls.length) throw new Error('No RPC URLs provided');
-  return {
-    type: ProviderType.CosmJsWasm,
-    provider: CosmWasmClient.connect(rpcUrls[0].http),
-  };
-}
-
-export function defaultCosmJsNativeProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: number | string,
-): CosmJsNativeProvider {
-  if (!rpcUrls.length) throw new Error('No RPC URLs provided');
-  return {
-    type: ProviderType.CosmJsNative,
-    provider: CosmosNativeProvider.connect(
-      rpcUrls.map((rpc) => rpc.http),
-      network,
-    ),
-  };
-}
-
-export function defaultStarknetJsProviderBuilder(
-  rpcUrls: RpcUrl[],
-): StarknetJsProvider {
-  assert(rpcUrls.length, 'No RPC URLs provided');
-  const { url, headers } = parseCustomRpcHeaders(rpcUrls[0].http);
-  const provider = new StarknetRpcProvider({
-    nodeUrl: url,
-    headers,
-  });
-  return { provider, type: ProviderType.Starknet };
-}
-
-export function defaultZKSyncProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: providers.Networkish,
-): ZKSyncProvider {
-  assert(rpcUrls.length, 'No RPC URLs provided');
-  const url = rpcUrls[0].http;
-  const provider = new ZKProvider(url, network);
-  return { type: ProviderType.ZkSync, provider };
-}
-
-export function defaultRadixProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: string | number,
-): RadixProvider {
-  assert(isNumeric(network), 'Radix requires a numeric network id');
-  const networkId = parseInt(network.toString(), 10);
-  const provider = new RadixSDKProvider({
-    rpcUrls: rpcUrls.map((rpc) => rpc.http),
-    networkId,
-  });
-  return { provider, type: ProviderType.Radix };
-}
-
-export function defaultAleoProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: string | number,
-): AleoProvider {
-  const provider = new AleoSDKProvider(
-    rpcUrls.map((rpc) => rpc.http),
-    network,
-  );
-  return { provider, type: ProviderType.Aleo };
-}
-
-/**
- * Returns an ethers-compatible TronJsonRpcProvider for use in MultiProvider.
- * This handles Tron's missing eth_getTransactionCount and returns the raw provider.
- */
-export function defaultTronEthersProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: number | string,
-): providers.Provider {
-  assert(rpcUrls.length > 0, 'At least one RPC URL required for Tron');
-  return new TronJsonRpcProvider(rpcUrls[0].http, network);
-}
-
-export function defaultTronProviderBuilder(
-  rpcUrls: RpcUrl[],
-  network: string | number,
-): TronProvider {
-  assert(rpcUrls.length > 0, 'At least one RPC URL required for Tron');
-  return {
-    provider: new TronJsonRpcProvider(rpcUrls[0].http, network),
-    type: ProviderType.Tron,
-  };
-}
-
-// Kept for backwards compatibility
-export function defaultProviderBuilder(
-  rpcUrls: RpcUrl[],
-  _network: number | string,
-): providers.Provider {
-  return defaultEthersV5ProviderBuilder(rpcUrls, _network).provider;
-}
-
-export function defaultZKProviderBuilder(
-  rpcUrls: RpcUrl[],
-  _network: number | string,
-): ZKProvider {
-  return defaultZKSyncProviderBuilder(rpcUrls, _network).provider;
-}
+  defaultCosmJsNativeProviderBuilder,
+  defaultCosmJsProviderBuilder,
+  defaultCosmJsWasmProviderBuilder,
+} from './builders/cosmos.js';
+import { defaultEthersV5ProviderBuilder } from './builders/ethersV5.js';
+import { defaultRadixProviderBuilder } from './builders/radix.js';
+import { defaultSolProviderBuilder } from './builders/solana.js';
+import { defaultStarknetJsProviderBuilder } from './builders/starknet.js';
+import { defaultTronProviderBuilder } from './builders/tron.js';
+import { defaultViemProviderBuilder } from './builders/viem.js';
+import { defaultZKSyncProviderBuilder } from './builders/zksync.js';
 
 export type ProviderBuilderMap = Record<
   ProviderType,
