+++ typescript/sdk/src/router/adapters/EvmRouterAdapter.ts
@@ -7,7 +7,7 @@ import {
 import { Address, Domain, bytes32ToAddress } from '@hyperlane-xyz/utils';
 
 import { BaseEvmAdapter } from '../../app/MultiProtocolApp.js';
-import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../../providers/ConfiguredMultiProtocolProvider.js';
 import { ChainName } from '../../types.js';
 
 import { IGasRouterAdapter, IRouterAdapter } from './types.js';
