+++ typescript/sdk/src/signers/radix/radix-toolkit.ts
@@ -5,7 +5,7 @@ import type {
 } from '@hyperlane-xyz/radix-sdk/runtime';
 import { RadixSigner } from '@hyperlane-xyz/radix-sdk/runtime';
 
-import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../../providers/ConfiguredMultiProtocolProvider.js';
 import { SendTransactionOptions } from '../../providers/MultiProvider.js';
 import { RadixTransaction } from '../../providers/ProviderType.js';
 import { ChainName } from '../../types.js';
