+++ typescript/sdk/src/token/Token.ts
@@ -12,7 +12,7 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import { ChainMetadata } from '../metadata/chainMetadataTypes.js';
-import { MultiProtocolProvider } from '../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../providers/ConfiguredMultiProtocolProvider.js';
 import { ChainName } from '../types.js';
 import { isStarknetFeeToken } from '../utils/starknet.js';
