+++ typescript/sdk/src/token/adapters/CosmosModuleTokenAdapter.ts
@@ -12,7 +12,7 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import { BaseCosmNativeAdapter } from '../../app/MultiProtocolApp.js';
-import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../../providers/ConfiguredMultiProtocolProvider.js';
 import { ChainName } from '../../types.js';
 import { PROTOCOL_TO_DEFAULT_NATIVE_TOKEN } from '../nativeTokenMetadata.js';
 import { TokenMetadata } from '../types.js';
