+++ typescript/sdk/src/token/adapters/CosmosTokenAdapter.ts
@@ -3,7 +3,7 @@ import { MsgSendEncodeObject, MsgTransferEncodeObject } from '@cosmjs/stargate';
 import { Address, Domain, assert } from '@hyperlane-xyz/utils';
 
 import { BaseCosmosAdapter } from '../../app/MultiProtocolApp.js';
-import { MultiProtocolProvider } from '../../providers/MultiProtocolProvider.js';
+import type { ConfiguredMultiProtocolProvider as MultiProtocolProvider } from '../../providers/ConfiguredMultiProtocolProvider.js';
 import { ChainName } from '../../types.js';
 import { TokenMetadata } from '../types.js';
